fixes
playing around w/ colors... still temp (just got sick of grey on grey)

-footer fix:
problem: was not working well with snapping panels.
fix: footer to be included on EACH individual page, as one of the scroll/snap components.  because it's not a static element, having it outside the router was causing weird behavior.  Works now very well