playing around with scrollProgress
So right now, the "Main" is using REF to capture the scroll position.

This can be passed as a property to any components (so in the background, this could animate certain things)

I've been just playing around with moving/resizing/changing colors, and this 100% can be used for the Landing page.

So where it's at now looks busy but I just left in as like a demo purpose for how to use it (still lots of chatgpt but took a little while to get it actually working)