global css + semantic changes
-using project panels to test this (because I've never done this before) using var(--project-panel-background) for example, setting this all in App.css, then we don't have to pick colors in child css files.

Semantic naming: looked into this a bit, and instead of "divs" everywhere, we should be using "header" for titles, "figure" for images, etc.  Again, in App.css we can set their style back to default, and in our files, gradually change <div> everywhere to these.  Might require some fixes throughout but long term it's going to be easier to work with.

-pulled down box shadow/border effect for project panels, based on Bobby's "looks like word perfect" comment lol.  I think boxing the elements off too much is not the look they're after.  so more floating elements in general.